debug make hex

+++ csf_assign01/fixedpoint.c
@@ -358,8 +358,8 @@ char *fixedpoint_format_as_hex(Fixedpoint val) {
     printf("\n%s\n",s);
   }
   printf("NEG CHECKED\n");
-  int ptr = (1<<63);
-  int back_shift = (1<<60);
+  uint64_t ptr = (1<<63);
+  uint64_t back_shift = (1<<60);
   int leading_zero = 1;
   for(int i = 0; i < 16; i++) { //67
     int hex = 0;
